Shifted AsciiDataLookup in utilities.h
