+++ include/aspect/boundary_temperature/ascii_data.h
@@ -23,8 +23,8 @@
 #define __aspect__boundary_temperature_ascii_data_h
 
 #include <aspect/boundary_temperature/interface.h>
-#include <aspect/velocity_boundary_conditions/ascii_data.h>
 #include <aspect/simulator_access.h>
+#include <aspect/utilities.h>
 
 
 namespace aspect
@@ -189,7 +189,7 @@ namespace aspect
          * data we get from text files.
          */
         std::map<types::boundary_id,
-                  std_cxx11::shared_ptr<::aspect::VelocityBoundaryConditions::internal::AsciiDataLookup<dim,dim-1> > > lookups;
+                  std_cxx11::shared_ptr<::aspect::Utilities::AsciiDataLookup<dim,dim-1> > > lookups;
 
         /**
          * Handles the update of the data in lookup.
