+++ source/compositional_initial_conditions/ascii_data.cc
@@ -22,7 +22,6 @@
 #include <aspect/global.h>
 #include <aspect/compositional_initial_conditions/ascii_data.h>
 
-#include <aspect/utilities.h>
 
 namespace aspect
 {
@@ -44,8 +43,7 @@ namespace aspect
       for (unsigned int i = 0; i < dim; i++)
         boundary_dimensions[i] = i;
 
-      lookup.reset(new VelocityBoundaryConditions::internal::AsciiDataLookup<dim,dim>    (data_points,
-                                                                                          this->get_geometry_model(),
+      lookup.reset(new Utilities::AsciiDataLookup<dim,dim>    (this->get_geometry_model(),
                                                                                           this->n_compositional_fields(),
                                                                                           scale_factor));
 
@@ -141,9 +139,6 @@ namespace aspect
           data_file_name    = prm.get ("Data file name");
 
           scale_factor      = prm.get_double ("Scale factor");
-          data_points[0]    = prm.get_integer ("Number of x grid points");
-          data_points[1]    = prm.get_integer ("Number of y grid points");
-          data_points[2]    = prm.get_integer ("Number of z grid points");
         }
         prm.leave_subsection();
       }
@@ -163,7 +158,11 @@ namespace aspect
                                                      "Implementation of a model in which the initial "
                                                      "composition is derived from files containing data "
                                                      "in ascii format. Note the required format of the "
-                                                     "input data: The order of the columns "
+                                                     "input data: The first lines may contain any number of comments"
+                                                     "if they begin with '#', but one of these lines needs to"
+                                                     "contain the number of grid points in each dimension as"
+                                                     "for example '# POINTS: 3 3'."
+                                                     "The order of the data columns "
                                                      "has to be 'x', 'y', 'composition_1', 'composition_2', "
                                                      "etc. in a 2d model and 'x', 'y', 'z', 'composition_1', "
                                                      "'composition_2', etc. in a 3d model, according "
